chore(devops): implement continuous delivery
Since Zinnia is now an UPM package it should automatically be
released as one. This change adds automatic creation of UPM packages
for Zinnia, including automatic SemVer-styled versioning based on
the commit messages. A release is both uploaded to the npm registry
at npmjs.com as well as an GitHub release (archived .zip).