chore(package): use Malimbe as dependency package
The compiled output of the dependency Malimbe was copied directly
into this project. Keeping binaries of dependencies in the project
only is necessary in environments that don't offer the idea of a
package and a package manager. With Unity nowadays coming with the
Unity Package Manager (UPM) Zinnia can now become a package,
referencing its needed dependency Malimbe as a package dependency.

BREAKING CHANGE: Zinnia is now a UPM package and does no longer
directly include Malimbe, instead referencing it as a package
dependency. Consumers of Zinnia will have to follow the added steps
in the README to include Zinnia in their projects.